audite/requirements.txt
Flask
Flask-Cors
gunicorn
Flask-SQLAlchemy
Flask-WTF
email-validator
python-dotenv
urllib3
werkzeug
fastapi
uvicorn
sqlalchemy
pydantic
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
PyJWT
requests